+++ src/main/java/com/inconcert/domain/post/service/EditService.java
@@ -52,6 +52,12 @@ public class EditService {
 
         postDto.setThumbnailUrl(extractURL(postDto.getContent()));
 
+        // 동행을 제외한 카테고리들은 모집인원, 마감 날짜 제거
+        if(!category.equals("match")){
+            postDto.setMatchCount(0);
+            postDto.setEndDate(null);
+        }
+
         // 새로운 레포지토리에 저장
         Post updatedPost = Post.builder()
                 .id(currentPost.getId())
