+++ src/main/java/com/inconcert/domain/post/service/InfoService.java
@@ -1,16 +1,10 @@
 package com.inconcert.domain.post.service;
 
-import com.inconcert.domain.category.entity.Category;
-import com.inconcert.domain.category.entity.PostCategory;
-import com.inconcert.domain.category.repository.CategoryRepository;
-import com.inconcert.domain.category.repository.PostCategoryRepository;
 import com.inconcert.domain.crawling.service.PerformanceService;
-import com.inconcert.domain.notification.service.NotificationService;
 import com.inconcert.domain.post.dto.PostDto;
 import com.inconcert.domain.post.entity.Post;
 import com.inconcert.domain.post.repository.InfoRepository;
 import com.inconcert.domain.post.util.DateUtil;
-import com.inconcert.domain.user.service.UserService;
 import com.inconcert.global.exception.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,15 +15,11 @@ import java.util.*;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class InfoService {
     private final InfoRepository infoRepository;
-    private final PostCategoryRepository postCategoryRepository;
-    private final CategoryRepository categoryRepository;
-    private final UserService userService;
     private final PerformanceService performanceService;
-    private final NotificationService notificationService;
 
-    @Transactional(readOnly = true)
     public List<PostDto> getAllInfoPostsByPostCategory(String postCategoryTitle) {
         List<Post> posts = switch (postCategoryTitle) {
             case "musical" -> infoRepository.findPostsByPostCategoryTitle("musical");
@@ -41,12 +31,10 @@ public class InfoService {
         return getPostDtos(posts);
     }
 
-    @Transactional(readOnly = true)
     public Post getPostByPostId(Long postId) {
         Optional<Post> post = infoRepository.findById(postId);
         return post.orElseThrow(() -> new PostNotFoundException(ExceptionMessage.POST_NOT_FOUND.getMessage()));
     }
-
 
     // postId를 가지고 게시물을 조회해서 postDto을 리턴해주는 메소드
     @Transactional
@@ -77,7 +65,6 @@ public class InfoService {
                 .build();
     }
 
-    @Transactional(readOnly = true)
     public List<PostDto> findByKeywordAndFilters(String postCategoryTitle, String keyword, String period, String type) {
         LocalDateTime startDate = DateUtil.getStartDate(period);
         LocalDateTime endDate = DateUtil.getCurrentDate();
@@ -89,47 +76,15 @@ public class InfoService {
     }
 
     @Transactional
-    public Post save(PostDto postDto){
-
-        // 게시물 작성 폼에서 가져온 postCategory 제목으로 조회해서 PostCategory 리스트 생성
-        List<PostCategory> postCategories = postCategoryRepository.findByTitle(postDto.getPostCategoryTitle());
-
-        // 게시물 작성 폼에서 가져온 Category 제목으로 조회해서 Category 객체 생성
-        Category category = categoryRepository.findByTitle(postDto.getCategoryTitle())
-                .orElseThrow(() -> new CategoryNotFoundException(ExceptionMessage.CATEGORY_NOT_FOUND.getMessage()));
-
-        // 적절한 PostCategory 찾기
-        PostCategory postCategory = postCategories.stream()
-                .filter(pc -> pc.getCategory().equals(category))
-                .findFirst()
-                .orElseThrow(() -> new PostCategoryNotFoundException(ExceptionMessage.POST_CATEGORY_COMBINATION_NOT_FOUND.getMessage()));
-
-        // 생성한 Category를 builder를 통해 연관관계 주입
-        PostCategory updatedPostCategory = postCategory.builder()
-                .id(postCategory.getId())
-                .title(postCategory.getTitle())
-                .category(category)
-                .build();
-
-        postDto.setUser(userService.getAuthenticatedUser()
-                .orElseThrow(() -> new UserNotFoundException(ExceptionMessage.USER_NOT_FOUND.getMessage())));
-
-        // 주입된 PostCategory를 Post에 저장
-        Post post = PostDto.toEntity(postDto, updatedPostCategory);
-
-        Post savePost = infoRepository.save(post);
-
-        // 알림 생성 로직 추가
-        notificationService.keywordsNotification(post);
-
-        return savePost;
-    }
-
-    @Transactional
     public void deletePost(Long postId) {
         Post post = infoRepository.findById(postId)
                 .orElseThrow(() -> new PostNotFoundException(ExceptionMessage.POST_NOT_FOUND.getMessage()));
         infoRepository.delete(post);
+    }
+
+    @Transactional
+    public void crawlAndSavePosts(String type) {
+        performanceService.crawlPerformances(type);
     }
 
     private static List<PostDto> getPostDtos(List<Post> posts) {
@@ -151,10 +106,4 @@ public class InfoService {
         }
         return postDtos;
     }
-
-    @Transactional
-    public void crawlAndSavePosts(String type) {
-        performanceService.crawlPerformances(type);
-    }
-
 }
