+++ src/main/resources/templates/board/editform.html
@@ -6,6 +6,39 @@
     <link rel="stylesheet" href="/css/board/writeform.css">
     <link href="https://cdn.quilljs.com/1.3.6/quill.snow.css" rel="stylesheet">
     <script src="https://cdn.quilljs.com/1.3.6/quill.min.js"></script>
+    <script>
+        function validateSearch() {
+            let keyword = document.getElementById("search-input").value.trim();
+            if (keyword.length < 2) {
+                alert("검색어를 2자 이상 입력하세요.");
+                return false;
+            }
+            return true;
+        }
+
+        function toggleMatchAndDateFields() {
+            const categoryTitle = document.getElementById("categoryTitle");
+            const matchCount = document.getElementById("matchCount");
+            const endDate = document.getElementById("endDate");
+            const dateContainer = document.querySelector(".date-container");
+
+            if (categoryTitle.value === "match") {
+                matchCount.style.display = "block";
+                dateContainer.style.display = "flex";
+            } else {
+                matchCount.style.display = "none";
+                dateContainer.style.display = "none";
+            }
+        }
+
+        document.addEventListener("DOMContentLoaded", function() {
+            const categoryTitle = document.getElementById("categoryTitle");
+            categoryTitle.addEventListener("change", toggleMatchAndDateFields);
+
+            // 페이지 로드 시 초기 상태 설정
+            toggleMatchAndDateFields();
+        });
+    </script>
 </head>
 <body>
 <div class="container">
@@ -33,7 +66,7 @@
             </div>
 
             <div class="dropdowns">
-                <select id="matchCount" name="matchCount" th:field="${post.matchCount}" required>
+                <select id="matchCount" name="matchCount" th:field="${post.matchCount}">
                     <option value="1">1</option>
                     <option value="2">2</option>
                     <option value="3">3</option>
@@ -41,7 +74,7 @@
                 </select>
                 <div class="date-container">
                     <label for="endDate">종료 날짜 지정</label>
-                    <input type="date" id="endDate" name="endDate" th:value="${post.endDate}" required>
+                    <input type="date" id="endDate" name="endDate" th:value="${post.endDate}">
                 </div>
             </div>
 
@@ -111,14 +144,17 @@
     document.getElementById('editForm').addEventListener('submit', function(event) {
         document.getElementById('content').value = quill.root.innerHTML;
 
+        var categoryTitle = document.getElementById('categoryTitle');
+        var postCategoryTitle = document.getElementById('postCategoryTitle');
         var matchCount = document.getElementById('matchCount');
         var endDate = document.getElementById('endDate');
         var title = document.getElementById('title');
         var content = document.getElementById('content');
 
         if (
-            matchCount.value === 'default' ||
-            !endDate.value ||
+            categoryTitle.value === 'default' ||
+            postCategoryTitle.value === 'default' ||
+            (categoryTitle.value === 'match' && (matchCount.value === 'default' || !endDate.value)) ||
             !title.value.trim() ||
             !content.value.trim()
         ) {
